Move context code to a separate package within the segment package
